requirements.txt
Flask
tensorflow
python-dotenv
requests